Add documentation and tests for GaussianLaw.
The class now has documentation on its methods, and there is a JUnit test class
that can be used to output a list of test values from this distribution.

In addition, the common functionality of UniformLawTest and GaussianLawTest was
moved to a superclass (LawTest).  This class can be used to implement any
behavior that should be tested for all Law subclasses.  Here, it is being used
to add a method that outputs 10,000 values from the distribution.