Add unit tests to check the min/max range values.
The unit tests for the Law subclasses will now check whether or not the minimum
value (0) and maximum value (upperBound - 1) can be generated by the
distribution.

Because of the random nature of the laws, the tests will try to generate the
values for four seconds.  If they cannot generate a value in that time, they
will consider it a failure.