Deprecate the Task(TaskSession) constructor.
The constructor was (a) not being used with any value except null and (b)
mostly useless since Tasks need to be added to a TaskSession, and the value is
changed when they are added to a TaskSession.

Moving forward, Tasks should be created using the no-arg constructor.  The
relevant subclasses have been updated with this change.