Simply the implementation of Event's compareTo.
The code in Event's implemention of compareTo had an extra if statement that
wasn't necessary.  Removing it does not affect the behavior of the method.

Here is a simple proof of why the behavior is the same.

First, assume that timestamp == e.timestamp.  In this case, both possible code
paths will return a value of zero.  Thus, always taking the second code path
produces the correct behavior.

Next, assume that timestamp != e.timestamp.  In this case, we are guaranteed to
take the second code path, which is the one that we keep.

Thus, the first if statement is unnecessary code that doesn't change the
behavior and makes the implementation slightly more confusing.