Add documentation to the Message class.
The unused "size" and "delay" fields were also removed.  The only value ever
stored in the size field (and never read) was zero.  The delay field was
provided with a real value in certain circumstances, but its value was also
never used.  (If/when these are needed again, it shouldn't be too difficult to
add them.)