Add tests for the Cache class (and fix a bug).
Unit test were added for the Cache class.  In addition, a bug that I
introduced by changing an "Integer" to an "int" was fixed thanks to the
unit tests.

I'll probably continue adding unit tests to make sure everything is
working as a I intend.  It's also a good way to ensure that we are on
the same page with how everything should work.