Update Message/Network to use MessageReceiver.
The code used to send messages was assuming that the the destination was
always a Node.  While this is true in the current version of the code,
it won't necessarily always be that way.  In addition, the
MessageReceiver class was already in place to provide the generalized
idea of an object capable of receiving Message.

The code to send messages will now allow them to be delivered to any
object that implements the MessageReceiver interface.