Initial changes to fix Tasks/TaskSessions.
There were problems were TaskSessions would not have all of their tasks
properly executed.  (This was visible in the ConsistencyExample class.)
In addition, there was logic implemented on a per-Task subclass basis
that should have been applied universally to all Task execution logic.

Therefore, the system was changed slightly to put more reliance on the
TaskProcessor class so that the behavior can be easily shared among all
of its current and future subclasses.  In addition, each Task is now
much more self-contained, and it is the responsibility of the Task to
trigger the execution of the next Task.  (This happens as a result of
the Task reporting its completion.)