Add documentation to VM.  BREAKING CHANGES.
In addition, the VM class now inherits directly from Node instead of
ServerNode.  These changes break a lot of functionality, but that
functionality will be fixed in future commits as the ServerNode class is
phased out.

The commitTask() method was also renamed to commitDiskTask() since it
only applies to DiskTasks.