Remove run and runClientSim modes from Simizer.
These two modes rely on the presence of Policy objects and some sort of
a load balancer.  This functionality has been removed from the core of
the Simizer framework.  Therefore, they can't possibly function as
intended within the confines of the framework.

In addition, I believe their behavior is reliant upon the old
architecture of the system.  I think that the behavior should either

  (a) be moved into the examples framework where the policies and load
      balancing application are located

  (b) be updated to be able to more generically run a simulation against
      some particular application.  This would allow them to be used for
      the load balancing case and also any other custom client cases.