Offer both operation modes for ClientNode.
The documentation suggests that the ClientNode should run for some
amount of time (defined by its durationLaw).

However, the code for the clients was sending a fixed number of requests
before finishing.

Both modes are now possible, through the use of two different
constructors.  Once a client is created in one mode, it cannot be
changed to another mode.