Automatically assign Node IDs.
(This also removes the Network parameter from the Node constructor.)

Because the concept of Node IDs introduces another mechanism for
addressing Nodes, and because it is potentially error-prone to manually
assign Node IDs, these IDs are now assigned internally by the framework.

This also removes the need for client code to directly worry about Node
IDs.  They can use the Java objects, retrieving a Node ID only when it
is necessary.  For example, if they want to know the Node ID for a node
in their simulation, they should add code to print out the Node ID
during the simulation.