Automatically assign Request IDs.
These were being automatically assigned in most situations.  The times
were they were still being manually assigned were using garbage values.
To remove the need to explain why a Request ID needs to be assigned,
that parameter has been removed from the constructor.

The one exception is for the CAWA policy, but that is part of code that
still needs to be rewritten for the updated version of the framework.