Fix template loading in RequestFactory.
The template loading method is now an instance method as opposed to a
static method.  In addition, its name was changed to refer to templates
instead of requests.  This will help users to understand the difference
between Requests and Request Templates.

The constructor was also changed so that the internal data structure
does not need to be supplied to the constructor.