Clean up Request (constructors and properties).
The suggested constructor for Request was pretty confusing and had a lot
of values that were either (a) not used or (b) set to some sort of
default value.  To avoid problems and confusion, the constructor has
been simplified.

In addition, there was a somewhat hidden notion in the Request of
assigning a target application, "type," and parameters.  These can be
thought of as server/host, action, and parameters.  The constructor
makes this sort of priority, ordering, and naming much more obvious.

As an example, consider the URL:

    http://example.com/definition/find?word=programming

In this URL, the "example.com" corresponds to the application, the
"definition/find" is the action, and the "word=programming" is the
parameter.

Some of the extra methods to retrieve the same values under different
names were removed.  In addition, some of the setters were deprecated in
favor of setting the values upon creation.  The defining "URL"
(application, action, parameters) should not be able to change once a
Request has been created and sent.