Improve the formatting of the output.
The output is rather difficult to decipher, and there are a few
application-specific values that don't make sense in all contexts.

Therefore, I added a more generic way of outputing the results of the
simulation.  This can allow each usage to define its own custom output,
but it also allows for more readable output when running simulations.