Add customizable fields to Request.
There were a few variables in the Request class that were being used to
track information related to the load balancer or particular policies.
In order to reduce the dependence of Request on these specific
applications, a generic "custom fields" feature was added to the class.

While there are still certain dependencies within the Request class on
these values, they should be able to disappear as the policies and
printing logic is updated.