Make the Law constructors easier to understand.
When I was writing the first tutorial, I needed to write about how to
create a GaussianLaw for the Network.  In the code, the "upper bound"
was being specified for the Gaussian law.  The concept of an upper bound
didn't make sense in the situation, nor was it easy to explain the
Gaussian law that would be created as a result of the particular
constructor.  Therefore, the constructors for each of the Law classes
has been updated to be as logical as possible for that particular
distribution.

As part of this change, the upper bound was removed by default, though
it can still be added to particular Law instances if it is necessary.
This may be useful for the ClientNodes' requestLaw, but the existing
tutorial didn't require it since it uses a ConstantLaw there.  The
inclusive lower bound of zero still remains.

A few remaining unit tests were added.

Finally, comments were added to the remaining Law classes, and the
implementations of ExponentialLaw and ZipfLaw were updated to be easier
to understand.