Make it obvious that a TaskSession is in use.
In an effort to make the code easier to read and understand, as well as
to facilitate the creation of the tutorial where the Application class
is described, the various actions that can be chained together have been
moved into a separate class.  Instance of this class (which is an inner
class of VM) can be passed to the handle() and init() methods to make
the vocabulary of the code easier to understand.