Fix the parameters/query string of Requests.
The "params" used throughout the application were confusing and a bit
difficult to understand.  The idea seemed to be to simulate the query
string of the URL, but it didn't follow that entirely.

There were parts of the code that tokenized it and took every other
value, and other parts that depended on a very specific formulation.  In
an effort to make the field easier to understand and use, it now
supports a lot more of the URL query string concept.  Values are
accessed by they keys, and there is some default support for passing a
list of resources.