Make ProcTask accept non-neg instruction counts.
There was a bug where ProcTask objects were being initialized with
negative instruction counts.  In order to remedy the problem, an
exception is thrown if an instruction contains a negative number of
instructions.

This makes the problem easier to debug because it throws the exception
when the problem is created, not later in the simulation when the
problem is used.