Remove the unused ResourceFactory parameter.
There was an unused "max" parameter in the Resource factory that made it
a bit confusing.  In order to simplify the implementation, the parameter
was removed.

In addition, a simple one-argument constructor was added to allow
applications to avoid specifying a default number of resources (if that
functionality is not needed in an application).