Remove code that throws an unreachable exception.
I think that the code must have changed at some point.  If a value was
null, an exception would be thrown.  However, that particular value is
both final and set in the constructor.  Therefore, it will never be
null.  The check (and, as a result, the "throws" clause) was removed.

This also means that the simulation can be run without using a try-catch
block or a throws clause.