Make the ClientGenerator easier to understand.
Rather than reusing ClientNode instances (which results in their IDs
being reused as well), we just create new instances.  This will also
make it easier to track the history of particular ClientNodes, if there
is a user who wants that sort of functionality.

In addition, the "simultaneous active ClientNodes" parameter is
implemented more accurately and includes better documentation.