Return copies of Resources from StorageElement.
When reading Resources, we will now return copies of those Resources.
This means that the client application cannot accidentally modify the
contents of a Resource on the disk.  In order to modify the contents of
the disk, it is necessary to call the write(), modify(), or delete()
methods.  This is a better representation of how real systems work.
Changing something in RAM does not automatically change that same thing
in the file on the disk.  The changes must be explicitly written to
remain consistent.